# src/swagger/docs.notes.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Note App API",
    "version": "1.0.0",
    "description": "API documentation for Note App"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
    "/notes": {
      "get": {
        "tags": ["Notes"],
        "summary": "Get all notes",
        "operationId": "getNotes",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number of notes",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search term for note title or content",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved notes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer"
                    },
                    "currentPage": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Note"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Notes"],
        "summary": "Create a new note",
        "operationId": "createNote",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Note"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Note created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          }
        }
      }
    },
    "/notes/{id}": {
      "get": {
        "tags": ["Notes"],
        "summary": "Get a specific note by ID",
        "operationId": "getNote",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the note",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "404": {
            "description": "Note not found"
          }
        }
      },
      "put": {
        "tags": ["Notes"],
        "summary": "Update a specific note",
        "operationId": "updateNote",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the note to update",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Note"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "404": {
            "description": "Note not found"
          }
        }
      },
      "delete": {
        "tags": ["Notes"],
        "summary": "Delete a note",
        "operationId": "deleteNote",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the note to delete",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note deleted successfully"
          },
          "404": {
            "description": "Note not found"
          }
        }
      },
      "patch": {
        "tags": ["Notes"],
        "summary": "Restore a soft-deleted note",
        "operationId": "restoreNote",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the note to restore",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note restored successfully"
          },
          "404": {
            "description": "Note not found"
          }
        }
      }
    },
    "/notes/{id}/pin": {
      "patch": {
        "tags": ["Notes"],
        "summary": "Toggle the pin status of a note",
        "operationId": "togglePin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the note to pin/unpin",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pin status updated"
          },
          "404": {
            "description": "Note not found"
          }
        }
      }
    },
    "/notes/{id}/archive": {
      "patch": {
        "tags": ["Notes"],
        "summary": "Toggle the archive status of a note",
        "operationId": "toggleArchive",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the note to archive/unarchive",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Archive status updated"
          },
          "404": {
            "description": "Note not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Note": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Note ID"
          },
          "title": {
            "type": "string",
            "description": "Title of the note"
          },
          "content": {
            "type": "string",
            "description": "Content of the note"
          },
          "is_pinned": {
            "type": "boolean",
            "description": "Pin status of the note"
          },
          "is_archived": {
            "type": "boolean",
            "description": "Archive status of the note"
          },
          "author_id": {
            "type": "integer",
            "description": "User ID who created the note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      }
    }
  }
}